+++ src/main/java/com/refinedmods/refinedstorage/apiimpl/storage/disk/PortableFluidStorageDisk.java
@@ -13,6 +13,7 @@ import net.minecraftforge.fluids.FluidStack;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Collection;
+import java.util.UUID;
 
 public class PortableFluidStorageDisk implements IStorageDisk<FluidStack> {
     private final IStorageDisk<FluidStack> parent;
@@ -26,6 +27,12 @@ public class PortableFluidStorageDisk implements IStorageDisk<FluidStack> {
     @Override
     public int getCapacity() {
         return parent.getCapacity();
+    }
+
+    @Nullable
+    @Override
+    public UUID getOwner() {
+        return parent.getOwner();
     }
 
     @Override