+++ src/main/java/com/refinedmods/refinedstorage/block/FluidStorageBlock.java
@@ -48,7 +48,7 @@ public class FluidStorageBlock extends NetworkNodeBlock {
                 storage.setStorageId(stack.getTag().getUniqueId(FluidStorageNetworkNode.NBT_ID));
             }
 
-            storage.loadStorage();
+            storage.loadStorage(player instanceof PlayerEntity ? (PlayerEntity) player : null);
         }
 
         // Call this after loading the storage, so the network discovery can use the loaded storage.