+++ src/main/java/com/refinedmods/refinedstorage/setup/ServerSetup.java
@@ -1,8 +1,9 @@
 package com.refinedmods.refinedstorage.setup;
 
 import com.refinedmods.refinedstorage.RS;
-import com.refinedmods.refinedstorage.command.CreateDiskCommand;
-import com.refinedmods.refinedstorage.command.PatternDumpCommand;
+import com.refinedmods.refinedstorage.command.disk.CreateDiskCommand;
+import com.refinedmods.refinedstorage.command.disk.ListDiskCommand;
+import com.refinedmods.refinedstorage.command.pattern.PatternDumpCommand;
 import net.minecraft.command.Commands;
 import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -10,7 +11,9 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 public class ServerSetup {
     @SubscribeEvent
     public void onRegisterCommands(RegisterCommandsEvent e) {
-        e.getDispatcher().register(Commands.literal(RS.ID).then(PatternDumpCommand.register()));
-        e.getDispatcher().register(Commands.literal(RS.ID).then(CreateDiskCommand.register()));
+        e.getDispatcher().register(Commands.literal(RS.ID).then(Commands.literal("pattern").then(PatternDumpCommand.register())));
+        e.getDispatcher().register(Commands.literal(RS.ID).then(Commands.literal("disk")
+            .then(CreateDiskCommand.register())
+            .then(ListDiskCommand.register())));
     }
 }