+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Refined Storage Changelog
 
 ### 1.9.5
-- Re-implemented the /refinedstorage createdisk command (raoulvdberge)
+- Re-implemented the /refinedstorage disk create <player> <id> command (raoulvdberge)
+- Added the /refinedstorage disk list command (raoulvdberge)
+- Added the /refinedstorage disk list <player> command (raoulvdberge)
 - Added JEI ghost ingredient dragging support (raoulvdberge)
 
 ### 1.9.4